refactor: Extract zerver/lib/email_validation.py.
