+++ zerver/forms.py
@@ -16,6 +16,7 @@ from jinja2 import Markup as mark_safe
 
 from zerver.lib.actions import do_change_password, email_not_system_bot, \
     validate_email_not_already_in_realm
+from zerver.lib.email_validation import email_allowed_for_realm
 from zerver.lib.name_restrictions import is_reserved_subdomain, is_disposable_domain
 from zerver.lib.rate_limiter import RateLimited, get_rate_limit_result_from_request, \
     RateLimitedObject, rate_limit_entity
@@ -25,7 +26,7 @@ from zerver.lib.subdomains import get_subdomain, is_root_domain_available
 from zerver.lib.users import check_full_name
 from zerver.models import Realm, get_user_by_delivery_email, UserProfile, get_realm, \
     email_to_domain, \
-    email_allowed_for_realm, DisposableEmailError, DomainNotAllowedForRealmError, \
+    DisposableEmailError, DomainNotAllowedForRealmError, \
     EmailContainsPlusError
 from zproject.backends import email_auth_enabled, email_belongs_to_ldap, check_password_strength
